requirements.txt
cffi==1.17.1
chacha20poly1305==0.0.3
Cifradescifra==2.0.0
cryptography==44.0.0
pycparser==2.22
pycryptodome==3.21.0
python-dotenv==1.0.1
python-gettext==5.0
